Changed method of creating vector - Worked better
I changed the the method of the skill you sent me Trent, tried it on my
server and it worked fine!
If you don't like the change the revert the commit!
I also added a check if the player is holding the item before checking
if the hero doing the skill is the target.